+++ sandwich/CHANGELOG.md
@@ -8,9 +8,9 @@
 * Add command-line arguments for `sandwich-quickcheck`.
 * Add `--visibility-threshold`/`-v` option.
 * Fix an issue where TUI quit would hang in the presence of cleared results.
-* Improve `runSandwichWithCommandLineArgs` to better detect when the TUI formatter can be run.
 * Display number of tests run in output message.
 * Catch and log exceptions in the TUI event loop.
+* Change `runSandwichWithCommandLineArgs` to use the print formatter by default, since we can't figure out how to detect if we're running under `cabal test`, which redirects stdout.
 
 ## 0.1.0.8